First  version of ArduWatchOS software
The first version of the ArduWatchOS software. This version works only with ssd1331 displays. It has only been tested with the hm-10 bluetooth module. To upload the code download arduino ide, connect your arduino, copy and paste this code into the ide, select the proper COM port, and hit upload.